refactor(og): extract shared image helpers into lib/og-render
